requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
bcrypt==4.0.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.8
Django==4.2.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
orjson==3.10.5
pipenv==2023.12.0
platformdirs==4.2.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
setuptools==69.0.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.4.4
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uvicorn==0.30.1
virtualenv==20.25.0
watchfiles==0.22.0
websockets==12.0
